refactor(stellar): Move poll config to ServerConfig
