+++ .env.example
@@ -47,3 +47,14 @@ KEYSTORE_PASSPHRASE=<keystore_passphrase> # This is the passphrase for the keyst
 #   Note: Cloud metadata endpoints (169.254.169.254) are ALWAYS blocked
 #   Example: RPC_BLOCK_PRIVATE_IPS=true
 # RPC_BLOCK_PRIVATE_IPS=false
+
+# Stellar Status Poll Cadence
+# STELLAR_STATUS_CHECK_INITIAL_DELAY_SECONDS: Delay before the first status check of a new Stellar transaction (default: 2, valid: 0-60)
+# STELLAR_STATUS_CHECK_INITIAL_DELAY_SECONDS=2
+
+# STELLAR_STATUS_RETRY_INITIAL_MS: Initial delay between Stellar status checks in milliseconds (default: 2000, valid: 500-60000)
+# STELLAR_STATUS_RETRY_INITIAL_MS=2000
+
+# STELLAR_STATUS_RETRY_MAX_MS: Maximum delay between Stellar status checks in milliseconds (default: 3000, valid: 500-60000)
+#   Clamped up to STELLAR_STATUS_RETRY_INITIAL_MS if set lower
+# STELLAR_STATUS_RETRY_MAX_MS=3000
