+++ docs/configuration/index.mdx
@@ -114,6 +114,9 @@ This table lists the environment variables and their default values.
 | `BACKGROUND_WORKER_TRANSACTION_STATUS_CHECKER_CONCURRENCY`         | `50`                     | `<any positive number>`                                                     | Maximum number of concurrent generic/default transaction status check jobs that can be processed simultaneously. This worker handles Solana and any future networks that don’t have dedicated status checkers.                                                                       |
 | `BACKGROUND_WORKER_TRANSACTION_STATUS_CHECKER_EVM_CONCURRENCY`     | `100`                    | `<any positive number>`                                                     | Maximum number of concurrent EVM transaction status check invocations that can be processed simultaneously. EVM handles the highest volume (~75% of all status checks) and uses optimized retries (8-20s) to avoid triggering premature resubmission.                                |
 | `BACKGROUND_WORKER_TRANSACTION_STATUS_CHECKER_STELLAR_CONCURRENCY` | `50`                     | `<any positive number>`                                                     | Maximum number of concurrent Stellar transaction status checks that can be processed simultaneously. Stellar status checker uses fast retries (2-3s) optimized for Stellar’s faster block times.                                                                                     |
+| `STELLAR_STATUS_CHECK_INITIAL_DELAY_SECONDS`                       | `2`                      | `0-60`                                                                      | Delay in seconds before the first status check of a new Stellar transaction. Invalid values fall back to the default.                                                                                                                                                                |
+| `STELLAR_STATUS_RETRY_INITIAL_MS`                                  | `2000`                   | `500-60000`                                                                 | Initial delay in milliseconds between Stellar transaction status checks. Invalid values fall back to the default.                                                                                                                                                                    |
+| `STELLAR_STATUS_RETRY_MAX_MS`                                      | `3000`                   | `500-60000`                                                                 | Maximum delay in milliseconds between Stellar transaction status checks. Clamped up to `STELLAR_STATUS_RETRY_INITIAL_MS` if set lower. Invalid values fall back to the default.                                                                                                      |
 | `BACKGROUND_WORKER_NOTIFICATION_SENDER_CONCURRENCY`                | `30`                     | `<any positive number>`                                                     | Maximum number of concurrent notifications that can be processed simultaneously.                                                                                                                                                                                                     |
 | `BACKGROUND_WORKER_SOLANA_TOKEN_SWAP_REQUEST_CONCURRENCY`          | `10`                     | `<any positive number>`                                                     | Maximum number of concurrent Solana token swap requests that can be processed simultaneously. Low volume worker.                                                                                                                                                                     |
 | `BACKGROUND_WORKER_TRANSACTION_CLEANUP_CONCURRENCY`                | `1`                      | `<any positive number>`                                                     | Maximum number of concurrent transaction cleanup invocations that can be processed simultaneously. Defaults to 1 to avoid database conflicts.                                                                                                                                        |
